fastAPI/requirements.txt
fastapi
uvicorn
requests
opentelemetry-api
opentelemetry-sdk